+++ src/view.rs
@@ -13,8 +13,8 @@ use crate::bytes::is_subslice;
 use crate::erase_lifetime;
 use crate::{bytes::ByteOwner, Bytes};
 use std::any::Any;
-use std::sync::Weak;
-use std::{fmt::Debug, hash::Hash, ops::Deref, sync::Arc};
+use std::sync::{Arc, Weak};
+use std::{fmt::Debug, hash::Hash, ops::Deref};
 use zerocopy::{Immutable, IntoBytes, KnownLayout, TryCastError, TryFromBytes};
 
 /// Errors that can occur when constructing a [`View`].
@@ -228,13 +228,15 @@ impl<T: ?Sized + Immutable> View<T> {
     }
 
     /// Returns the owner of the View in an `Arc`.
-    pub fn downcast_to_owner<O>(self) -> Option<Arc<O>>
+    pub fn downcast_to_owner<O>(self) -> Result<Arc<O>, View<T>>
     where
         O: Send + Sync + 'static,
     {
-        let owner = self.owner;
-        let any: Arc<dyn Any + Send + Sync> = owner;
-        Arc::downcast::<O>(any).ok()
+        let owner_any: Arc<dyn Any + Send + Sync> = self.owner.clone();
+        match owner_any.downcast::<O>() {
+            Ok(owner) => Ok(owner),
+            Err(_) => Err(self),
+        }
     }
 
     /// Create a weak pointer.